+++ frontend/src/utils/dnd/index.js
@@ -1,0 +1,26 @@
+/**
+ * 文件名: index.js
+ * 组件: 拖拽工具模块入口文件
+ * 描述: 导出所有拖拽相关的组件和工具
+ * 作者: Jolly Chen
+ * 时间: 2024-11-20
+ * 版本: 1.2.0
+ * 许可证: Apache-2.0
+ */
+
+// 导出日志系统
+export { Logger, LOG_LEVEL, CURRENT_LOG_LEVEL } from './dnd-logger.js';
+
+// 导出可排序项组件
+export { createSortableItem } from './dnd-sortable.jsx';
+
+// 导出笔记拖拽上下文
+export { NoteDndContext } from './dnd-note-context.jsx';
+
+// 导出文件拖拽上下文
+export { FileDndContext } from './dnd-file-context.jsx';
+
+// 导出辅助函数
+export {
+  throttle
+} from './dnd-helpers.js';
